Update for new layout (November 2024)
The BreatheHR dashboard layout has changed quite substantially. This
updates the script to work with the new layout. This required updating
the selectors, waiting for dependent content to load (the new dashboard
seems to use hydration - the content isn't necessarily present on page
load), and updating the layout/styling to fit better with the new
BreatheHR dashboard styling

This includes a major version bump given it's made for a completely new
layout and won't work with the old one. The script has been working fine
for many people for quite some time, so I think it's fine to push it to
v1